Add files via upload
A company has made a couple of hackathons, many people contributed to those hackathons. The results
of those hackathons are stored in a file. Each line in the files of the results starts with the ID of the
hackathon (string without white spaces), then the name of the competitor (string without white spaces),
then the results of every task the competitor solved on the scale of 1..100 (integer), the competitor
solved at least one task!
get the competitor who got the max score before the first failing competitor (the score is the
average
GRADE 3
of all of the results of one competitor, and a competitor fails if he scored less than 40%)